requirements.txt
httpx~=0.24.1
python-telegram-bot
requests~=2.26.0
asyncio~=3.4.3
Pillow~=10.0.0
telegram~=0.0.1
pytz~=2023.3
trio~=0.22.2
watchdog~=3.0.0